+++ src/main/java/pl/edu/uws/pp/service/impl/InvoiceServiceImpl.java
@@ -1,9 +1,13 @@
 package pl.edu.uws.pp.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pl.edu.uws.pp.config.security.UserPrincipal;
 import pl.edu.uws.pp.domain.dto.invoice.*;
+import pl.edu.uws.pp.domain.enums.InvoiceStatus;
+import pl.edu.uws.pp.domain.enums.Role;
 import pl.edu.uws.pp.domain.mapper.InvoiceMapper;
 import pl.edu.uws.pp.exception.NotFoundException;
 import pl.edu.uws.pp.repository.ApartmentRepository;
@@ -17,9 +21,15 @@ public class InvoiceServiceImpl implements InvoiceService {
     private final ApartmentRepository apartmentRepository;
 
     @Override
-    public InvoiceShortResponse createInvoice(InvoiceRequest request) {
+    public InvoiceShortResponse createInvoice(InvoiceRequest request,
+                                              UserPrincipal principal) {
         var apartment = apartmentRepository.findById(request.apartmentId())
                 .orElseThrow(() -> new NotFoundException("nie znaleziono mieszkania"));
+        var user = principal.user();
+        if (user.getManagerProfile().isNotManagingApartment(apartment)){
+            throw new AccessDeniedException("Nie możesz wystawić opłaty dla tego mieszkania");
+        }
+
         var invoice = InvoiceMapper.fromInvoiceRequest(request, apartment);
         var savedInvoice = invoiceRepository.save(invoice);
 
@@ -27,9 +37,19 @@ public class InvoiceServiceImpl implements InvoiceService {
     }
 
     @Override
-    public InvoiceResponse getInvoiceInfo(Long id) {
+    public InvoiceResponse getInvoiceInfo(Long id,
+                                          UserPrincipal principal) {
         var invoice = invoiceRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("nie znaleziono opłaty"));
+        var user = principal.user();
+        if (user.isRoleEqualed(Role.BUILDING_MANAGER)
+            && user.getManagerProfile().isNotManagingApartment(invoice.getApartment())){
+            throw new AccessDeniedException("Nie masz dostępu do tej opłaty");
+        }
+        if (user.isRoleEqualed(Role.RESIDENT)
+            && ! user.getResidentProfile().isOwningApartment(invoice.getApartment())) {
+            throw new AccessDeniedException("Nie masz dostępu do tej opłaty");
+        }
 
         return InvoiceMapper.toInvoiceResponse(invoice);
     }
@@ -37,11 +57,19 @@ public class InvoiceServiceImpl implements InvoiceService {
     @Override
     @Transactional
     public InvoiceShortResponse editInvoice(Long id,
-                                            InvoiceEditRequest request) {
+                                            InvoiceEditRequest request,
+                                            UserPrincipal principal) {
         var invoice = invoiceRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Nie znaleziono opłaty"));
         var apartment = apartmentRepository.findById(request.apartmentId())
                 .orElseThrow(() -> new NotFoundException("Nie znaleziono mieszkania"));
+        var user = principal.user();
+        if (user.getManagerProfile().isNotManagingApartment(invoice.getApartment())) {
+            throw new AccessDeniedException("Nie masz dostępu do tej opłaty");
+        }
+        if (user.getManagerProfile().isNotManagingApartment(apartment)) {
+            throw new AccessDeniedException("Nie możesz wystawić opłaty na to mieszkanie");
+        }
 
         invoice.setApartment(apartment);
         invoice.setDescription(request.description());
@@ -54,11 +82,16 @@ public class InvoiceServiceImpl implements InvoiceService {
     @Override
     @Transactional
     public InvoiceShortResponse changeInvoiceStatus(Long id,
-                                                    InvoiceChangeStatusRequest request) {
+                                                    InvoiceChangeStatusRequest request,
+                                                    UserPrincipal principal) {
         var invoice = invoiceRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("nie znaleziono opłaty"));
+        var user = principal.user();
+        if (user.getManagerProfile().isNotManagingApartment(invoice.getApartment())) {
+            throw new AccessDeniedException("Nie masz dostępu d tego mieszkania");
+        }
 
-        if (!invoice.getStatus().canChangeTo(request.status()))
+        if (! invoice.getStatus().canChangeTo(request.status()))
             throw new IllegalStateException("Nie można zmienić statusu");
         invoice.setStatus(request.status());
 
@@ -69,6 +102,9 @@ public class InvoiceServiceImpl implements InvoiceService {
     public void deleteInvoice(Long id) {
         var invoice = invoiceRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Nie znaleziono opłaty"));
+        if (invoice.getStatus().equals(InvoiceStatus.PAID)){
+            throw new IllegalStateException("Nie możesz usunąć opłaconej opłaty");
+        }
         invoiceRepository.delete(invoice);
     }
 }
